+++ packages/core/types/term-size/index.d.ts
@@ -1,3 +1,0 @@
-declare module 'term-size' {
-  export default function termSize(): { columns: number; rows: number };
-}